# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sera Test API",
        "contact": {
            "name": "Ibnu Muntoha",
            "email": "ibnu.muntoha@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/external/register": {
            "post": {
                "tags": [
                    "Handle External API"
                ],
                "summary": "Handle register",
                "description": "Handle register",
                "operationId": "Handle register",
                "requestBody": {
                    "description": "Handle register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "eve.holt@reqres.in"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "cityslicka"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "azgZaZT7bCdInSAymsHgyVy0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/external/login": {
            "post": {
                "tags": [
                    "Handle External API"
                ],
                "summary": "Handle Login",
                "description": "Handle Login",
                "operationId": "Handle Login",
                "requestBody": {
                    "description": "Handle Login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "eve.holt@reqres.in"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "cityslicka"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "azgZaZT7bCdInSAymsHgyVy0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}